fixing variable type

+++ singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-int countNode = 0;
+size_t countNode = 0;
 while (h != (void *)0)
 {
 if (h->str == (void *)0)
